lcm_num =  int((lcm_num * a[x])/gcd(lcm_num,a[x]))
Change of the line 9 to avoid 'float' object cannot be interpreted as an integer

+++ between-two-sets.py
@@ -6,7 +6,8 @@ def getTotalX(a, b):
     gcd_num = b[0]
     if len(a) > 1:
         for x in range(1,len(a)):
-            lcm_num =  (lcm_num * a[x])/gcd(lcm_num,a[x])
+            # lcm_num =  (lcm_num * a[x])/gcd(lcm_num,a[x])
+            lcm_num =  int((lcm_num * a[x])/gcd(lcm_num,a[x]))
     if len(b) > 1:
         for x in range(1,len(b)):
             gcd_num = gcd(gcd_num,b[x])
